Fixed a bad interaction between read_completed and JavaProxy.
When a Ruby class extends from Java class, JRuby uses two objects internally to represent an object. One is a Java instance that extends from the specific Java class, the other is an instance of the JavaProxy type which is a Ruby side of the object.

read_completed cannot be invoked until we have both properly unmarshalled.